move stat api response types to stat_types file
