+++ static/js/ranking/ranking_page.tsx
@@ -22,9 +22,9 @@ import { defineMessages } from "react-intl";
 import { intl, LocalizedLink } from "../i18n/i18n";
 import { displayNameForPlaceType } from "../place/util";
 import { DEFAULT_POPULATION_DCID } from "../shared/constants";
-import { StatApiResponse } from "../shared/stat_types";
+import { GetStatSetResponse, StatApiResponse } from "../shared/stat_types";
 import { getStatsVarTitle } from "../shared/stats_var_titles";
-import { getPopulationDate, GetStatSetResponse } from "../tools/shared_util";
+import { getPopulationDate } from "../tools/shared_util";
 import { RankingHistogram } from "./ranking_histogram";
 import { RankingTable } from "./ranking_table";
 import { LocationRankData, RankInfo } from "./ranking_types";
